LVSP-61. Sync merged create and update in google-services
